[Addition] Add an additional picker to `find_program()`
Currently `find_program()` finds executables in cwd. This is great if the binary is locally compiled, but not so if it is in some other directory.

Also I am too lazy to enter the same path again and again. So hardcoded the damn thing.

Maybe it would make more sense to have it as an option in some later commit